Almost finished with DB schema

+++ CommonJPA/src/main/java/ru/yanes/autoprom/dao/ComplectationDAO.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.yanes.autoprom.entity.Complectation;
 
 @Repository
-public interface ComplectationDAO extends JpaRepository<Complectation, Integer> {
+public interface ComplectationDAO extends JpaRepository<Complectation, Long> {
 }
